First tests: check the mode of the mock object.
Set it to either record or play (not a metaphor related to distilling,
but that's fine, not everything needs to have just one metaphor that's
then painfully extended to include every facet of the software.
Looking at you, Homebrew.)

Also build a sample test class that satisfies all the requirements of
the role.