Record new mocks.
When recording mocks, talk to a live service and remember the distilled
request and distilled response. Test that this happens, and that the
response we get back are what we expect.

Monkey-patching this stuff was really hard, which is why there's
explicit for monkey-patching: I couldn't find a way of messing with the
symbol table or %Class::Method::Modifiers::MODIFIER_CACHE.
HARNESS_ACTIVE is set by prove and all good tests, so is safe to rely
upon (famous last words).