Fix overriding / monkey-patching LWP::UserAgent
First of all, override simple_request, which is documented, rather than
send_request, which isn't. This fits with how LWP::UserAgent::JSON
messes with the guts of LWP.

Secondly, don't monkey-patch the method we're overriding when we test
the ::JSON version of the code, because that code is the code in
LWP::UserAgent::JSON which reblesses objects but otherwise hands
control to LWP::UserAgent. Instead, monkey-patch LWP::UserAgent directly
and avoid modifying the intermediate code we wanted to leave alone.
This is possible because LWP::UserAgent::JSON overrides the method via
inheritance rather than a symbol table override.